# influxdata: support legacy decoding